:sparkle: Feature: Add unidirectional map
