+++ rust/libs/kvs/src/map/bidirectional_map.rs
@@ -23,7 +23,8 @@ use sled::{
 use tracing::instrument;
 use std::{borrow::Borrow, sync::{atomic::AtomicUsize, Arc}};
 
-use crate::{codec::Codec, error::Error, map::Map, types::{KeyType, ValueType}};
+use crate::map::{codec::Codec, error::Error, base::MapBase, types::{KeyType, ValueType}};
+
 
 const PRIMARY_TREE_NAME: &str = "uo";
 const SECONDARY_TREE_NAME: &str = "ou";
@@ -41,7 +42,7 @@ pub struct BidirectionalMap<K: KeyType, V: ValueType, C: Codec> {
     _marker: std::marker::PhantomData<(K, V)>,
 }
 
-impl<K: KeyType, V: ValueType, C: Codec> Map for BidirectionalMap<K, V, C> {
+impl<K: KeyType, V: ValueType, C: Codec> MapBase for BidirectionalMap<K, V, C> {
     type K = K;
     type V = V;
     type C = C;
